remi-frontend/requirements.txt
pywebview==3.1
numpy==1.17.4
requests==2.22.0
SoundFile==0.10.2
SpeechRecognition==3.8.1
PyAudio==0.2.11
